feat(app): create blank labware position check generic step screen and break out intro screen
