+++ app/src/assets/localization/en/labware_position_check.json
@@ -1,0 +1,11 @@
+{
+  "labware_position_check_title": "Labware Position Check",
+  "labware_position_check_overview": "Labware Position Check Overview",
+  "position_check_description": "Labware Position Check check jogs your pipette(s) to each labware on the deck to ensure the OT-2 will move accurately during a protocol run. You will use {{number_of_tips}} tip from the {{labware_name}} in Slot {{labware_slot}} to check this protocol’s labware.",
+  "position_check_description_plural": "Labware Position Check check jogs your pipette(s) to each labware on the deck to ensure the OT-2 will move accurately during a protocol run. You will use {{number_of_tips}} tips from the {{labware_name}} in Slot {{labware_slot}} to check this protocol’s labware.",
+  "start_position_check": "begin labware position check, move to slot {{initial_labware_slot}}",
+  "secondary_pipette_tipracks_section": "Check tipracks with {{secondary_mount}} Pipette",
+  "primary_pipette_tipracks_section": "Check tipracks with {{primary_mount}} Pipette",
+  "check_remaining_labware_with_primary_pipette_section": "Check remaining labware with {{primary_mount}} Pipette",
+  "return_tip_section": "Return tip"
+}
