+++ app/src/organisms/ProtocolSetup/LabwarePositionCheck/GenericStepScreen.tsx
@@ -1,0 +1,3 @@
+import * as React from 'react'
+
+export const GenericStepScreen = (): JSX.Element | null => <div></div>
